Split forecasts (canonical & standalone)
