+++ conf/taranis.ini
@@ -298,7 +298,7 @@ partition           = login
 # FORECASTS
 # =========
 
-[canonical_forecasts]
+[canonical_forecasts_arpege_global798c22]
 prefix              = canonical
 # profile (ntasks = ntasks/node)
 time                = 01:00:00
@@ -306,7 +306,15 @@ ntasks              = 32
 nnodes              = 4
 openmp              = 4
 
-[standalone_forecasts]
+[canonical_forecasts_arpege_apl_global798c22]
+prefix              = canonical
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 4
+openmp              = 4
+
+[standalone_forecast_alaro0_antwrp1300]
 prefix              = standalone
 pgd_source          = static
 surf_ic_source      = static
@@ -315,6 +323,57 @@ time                = 01:00:00
 ntasks              = 32
 nnodes              = 1
 openmp              = 4
+
+[standalone_forecast_alaro1_antwrp1300]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_arome_corsica2500]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_arpege_globaltst149c24]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_arpege_nosfx_globaltst149c24]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_ifs_global21]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
 
 # MIXS
 # ====
